Initial 3.0 refactor (broken)
Added some new classes like TypeMap to make updating the protocol easier and more readable for other to understand.
Items and blocks will now be given to the user as item and block definitions so that runtime IDs do not need to be handled however, it is still possible.
BedrockPacketHelper is now BedrockCodecHelper and will no longer be a singleton. Instead an instance will be created for each class so that the item and block definitions and can be passed so that serialisation does not depend on BedrockSession.
Many methods that were needlessly added to BedrockPacketHelper have been moved out to their own serialisers including the TypeMaps.
LevelEventType has now been spit up into two separate enums: ParticleType and LevelEvent which both implement the LevelEventType interface so both can be set as the event type in LevelEventPacket whilst ParticleType can be used in entity data for area effect clouds.